% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Abalone}
\alias{Abalone}
\title{Abalone Dataset}
\format{
An object of class \code{data.frame} with 24 rows and 14 columns.
}
\usage{
data(Abalone)
}
\description{
A interval-valued data set containing 24 units,
created from from the Abalone dataset (UCI Machine Learning Repository), after aggregating by sex and age.
}
\examples{
data(Abalone)
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
