% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acid_rain.int}
\alias{acid_rain.int}
\title{Acid Rain Pollution Indices Interval Dataset}
\format{
A data frame with 2 observations and 5 variables in Min-Max format:
\itemize{
    \item \code{state}: State name (character).
    \item \code{sulphate_l}, \code{sulphate_u}: Sulphate pollution index range
          (kg/hectares).
    \item \code{nitrate_l}, \code{nitrate_u}: Nitrate pollution index range
          (kg/hectares).
}
}
\usage{
data(acid_rain.int)
}
\description{
Interval-valued acid rain pollution indices for sulphates and nitrates
(kg/hectares) for 2 US states (Massachusetts and New York).
}
\examples{
data(acid_rain.int)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley.
Table 2.21.
}
\keyword{datasets}
\keyword{interval}
