% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Age_pyramids}
\alias{Age_pyramids}
\title{Age pyramids of all the countries of the World in 2014}
\format{
a \code{MatH} object, a matrix of distributions.
}
\source{
\itemize{
    \item United States Census Bureau \url{https://www.census.gov/data.html}
    \item \url{https://CRAN.R-project.org/package=HistDAWass}
}
}
\usage{
data(Age_pyramids)
}
\description{
The dataset contains a MatH (matrix of histogram-valued data) object, with three
hisogram-valued variables, the 5-years age (relative frequencies) distribution of all the population,
of the male and of the female population of 228 countries of the World. The  first row is the World data.
Thus it contains 229 rows(228 countries plus the World) and 3 variables: "Both.Sexes.Population",
"Male.Population", "Female.Population"
}
\examples{
data(Age_pyramids)
}
\keyword{datasets}
