% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Airline_flights}
\alias{Airline_flights}
\alias{Airline_flights2}
\title{Airline Flights Dataset}
\format{
An object of class \code{data.frame} with 16 rows and 17 columns.

An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16 rows and 6 columns.
}
\usage{
data(Airline_flights)

data(Airline_flights2)
}
\description{
Airline Flights Dataset.

Airline Flights Modal-Valued Dataset.
}
\examples{
data(Airline_flights)
data(Airline_flights2)
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.

Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
