% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_format_conversions.R
\name{ARRAY_to_iGAP}
\alias{ARRAY_to_iGAP}
\title{ARRAY to iGAP}
\usage{
ARRAY_to_iGAP(data)
}
\arguments{
\item{data}{A numeric array of dimension \code{[n, p, 2]} where
\code{[,,1]} stores minima and \code{[,,2]} stores maxima.}
}
\value{
A data.frame in iGAP format with comma-separated \code{"min,max"}
values.
}
\description{
Convert a 3-dimensional array \code{[n, p, 2]} to iGAP format
(data.frame with comma-separated interval values).
}
\examples{
x <- array(NA, dim = c(4, 3, 2))
x[,,1] <- matrix(c(1,2,3,4, 5,6,7,8, 9,10,11,12), nrow = 4)
x[,,2] <- matrix(c(3,5,6,7, 8,9,10,12, 13,15,16,18), nrow = 4)
dimnames(x) <- list(paste0("obs_", 1:4), c("V1","V2","V3"), c("min","max"))
igap <- ARRAY_to_iGAP(x)
igap
}
