% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_format_conversions.R
\name{ARRAY_to_MM}
\alias{ARRAY_to_MM}
\title{ARRAY to MM}
\usage{
ARRAY_to_MM(data)
}
\arguments{
\item{data}{A numeric array of dimension \code{[n, p, 2]} where
\code{[,,1]} stores minima and \code{[,,2]} stores maxima.}
}
\value{
A data.frame with \code{2p} columns (paired \code{_min}/\code{_max}).
}
\description{
Convert a 3-dimensional array \code{[n, p, 2]} to MM format
(data.frame with paired \code{_min}/\code{_max} columns).
}
\examples{
x <- array(NA, dim = c(4, 3, 2))
x[,,1] <- matrix(c(1,2,3,4, 5,6,7,8, 9,10,11,12), nrow = 4)
x[,,2] <- matrix(c(3,5,6,7, 8,9,10,12, 13,15,16,18), nrow = 4)
dimnames(x) <- list(paste0("obs_", 1:4), c("V1","V2","V3"), c("min","max"))
mm <- ARRAY_to_MM(x)
mm
}
