% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_format_conversions.R
\name{ARRAY_to_RSDA}
\alias{ARRAY_to_RSDA}
\title{ARRAY to RSDA}
\usage{
ARRAY_to_RSDA(data)
}
\arguments{
\item{data}{A numeric array of dimension \code{[n, p, 2]} where
\code{[,,1]} stores minima and \code{[,,2]} stores maxima.}
}
\value{
A symbolic_tbl with p symbolic_interval columns.
}
\description{
Convert a 3-dimensional array \code{[n, p, 2]} to RSDA format
(symbolic_tbl with symbolic_interval columns).
}
\examples{
x <- array(NA, dim = c(4, 3, 2))
x[,,1] <- matrix(c(1,2,3,4, 5,6,7,8, 9,10,11,12), nrow = 4)
x[,,2] <- matrix(c(3,5,6,7, 8,9,10,12, 13,15,16,18), nrow = 4)
dimnames(x) <- list(paste0("obs_", 1:4), c("V1","V2","V3"), c("min","max"))
rsda <- ARRAY_to_RSDA(x)
rsda
}
