% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bird.mix}
\alias{bird.mix}
\title{Bird Species Mixed Symbolic Dataset}
\format{
A symbolic data frame (\code{symbolic_tbl}) with 20 observations and 2 variables:
\itemize{
    \item \code{Density}: Feather density range (interval).
    \item \code{Size}: Body size range (cm, interval).
}
}
\usage{
data(bird.mix)
}
\description{
Interval-valued morphological measurements for 20 bird specimens.
Despite the \code{.mix} suffix, this dataset contains only
interval-valued variables (density and size).
}
\examples{
data(bird.mix)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley. Table 2.5.
}
\keyword{datasets}
\keyword{interval}
