% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bird_color_taxonomy.hist}
\alias{bird_color_taxonomy.hist}
\title{Bird Color Taxonomy Histogram Dataset}
\format{
A data frame with 20 observations and 4 variables:
\itemize{
    \item \code{density}: Histogram-valued feather density (up to 4 bins).
    \item \code{size}: Histogram-valued body size (2-bin).
    \item \code{tone}: Categorical tone (dark/light).
    \item \code{shade}: Distribution-valued shade (purple/red/white/yellow
    with fuzzy weights).
}
}
\source{
Billard, L. and Diday, E. (2007), Tables 6.9/6.14.
}
\usage{
data(bird_color_taxonomy.hist)
}
\description{
Mixed symbolic dataset of 20 bird observations with histogram-valued
feather density and body size, categorical tone, and distribution-valued
shade (fuzzy taxonomy). From Tables 6.9 and 6.14 of Billard and Diday
(2007).
}
\examples{
data(bird_color_taxonomy.hist)
}
\references{
Billard, L. and Diday, E. (2007). \emph{Symbolic Data Analysis:
Conceptual Statistics and Data Mining}. Wiley, Chichester. Tables 6.9
and 6.14.
}
\keyword{datasets}
\keyword{distribution}
\keyword{histogram}
\keyword{mixed}
