% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Blood}
\alias{Blood}
\title{Blood dataset for Histogram data analysis}
\format{
a \code{MatH} istance, 1 row per group.
}
\source{
\itemize{
    \item Billard L. and Diday E. (2006). Symbolic Data Analysis: Conceptual Statistics and Data Mining, Wiley.
    \item \url{https://CRAN.R-project.org/package=HistDAWass}
}
}
\usage{
data(Blood)
}
\description{
The dataset contains a MatH (matrix of histogram-valued data)
object This data set list 14 groups of patients described by 3 variables.
}
\examples{
data(Blood)
}
\keyword{datasets}
