% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blood_pressure.int}
\alias{blood_pressure.int}
\title{Blood Pressure Interval Dataset}
\format{
A symbolic data frame (\code{symbolic_tbl}) with 15 observations and
3 interval-valued variables:
\itemize{
    \item \code{Pulse_Rate}: Pulse rate range (beats per minute, interval).
    \item \code{Systolic_Pressure}: Systolic blood pressure range (mmHg, interval).
    \item \code{Diastolic_Pressure}: Diastolic blood pressure range (mmHg, interval).
}
}
\usage{
data(blood_pressure.int)
}
\description{
Interval-valued blood pressure and pulse rate measurements for
15 patient groups.
}
\examples{
data(blood_pressure.int)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley.
}
\keyword{datasets}
\keyword{interval}
