% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Blood_pressure}
\alias{Blood_pressure}
\title{Blood Pressure Interval-Valued Dataset}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 15 rows and 3 columns.
}
\usage{
data(Blood_pressure)
}
\description{
blood pressure Interval-Valued Dataset.
}
\examples{
data(Blood_pressure)
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
