% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bloodbrito}
\alias{Bloodbrito}
\title{Blood dataset from Brito P. for Histogram data analysis}
\format{
a \code{MatH} istance, 1 row per patient.
}
\source{
\itemize{
    \item Dias, S. and Brito P. Distribution and Symmetric Distribution Regression Model for Histogram-Valued Variables, ArXiv, arXiv:1303.6199 [stat.ME]
    \item \url{https://CRAN.R-project.org/package=HistDAWass}
}
}
\usage{
data(Bloodbrito)
}
\description{
The dataset contains a MatH (matrix of histogram-valued data) object.
This data set list 10 patients described by 2 variables.
}
\examples{
data(Bloodbrito)
}
\keyword{datasets}
