% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cardiological2_data}
\alias{Cardiological2_data}
\title{Cardiological data example}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 44 rows and 5 columns.
}
\source{
\url{https://CRAN.R-project.org/package=RSDA}
}
\usage{
data(Cardiological2_data)
}
\description{
Cardiological interval data example.
}
\references{
Billard L. and  Diday E. (2006).
Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
#' @examples
data(Cardiological2_data)
}
\keyword{datasets}
