% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{China_month}
\alias{China_month}
\title{A monthly climatic dataset of China}
\format{
a \code{MatH} object, a matrix of distributions.
}
\source{
\itemize{
    \item raw data are available here: \url{https://cdiac.ess-dive.lbl.gov/ftp/tr055/}
    \item \url{https://CRAN.R-project.org/package=HistDAWass}
}
}
\usage{
data(China_month)
}
\description{
A dataset with the distributions of some climatic variables collected for each
month in 60 stations of China. The collected variables are 168 i.e. 14 climatic
variables observed for 12 months. The 14 variables are the following:
mean station pressure (mb), mean temperature, mean maximum temperature,
mean minimum temperature, total precipitation (mm), sunshine duration (h),
mean cloud amount (percentage of sky cover), mean relative humidity ( mean wind
speed (m/s), dominant wind frequency ( extreme minimum temperature ).
Use the command \code{ get.MatH.main.info(China_Month) } for rapid info.
}
\examples{
data(China_month)
}
\keyword{datasets}
