% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChinaTemp.int}
\alias{ChinaTemp.int}
\title{China Temperatures Interval Dataset}
\format{
A data frame containing 899 observations on 5 variables, the first four with the temperatures by quarter in the 899 stations-year
combinations, the last one a factor indicating the geographic region of each station.
}
\source{
\url{https://CRAN.R-project.org/package=MAINT.Data}
}
\usage{
data(ChinaTemp.int)
}
\description{
China Temperatures Interval Dataset generated from ChinaTemp dataset.
This data set consist of the intervals of observed temperatures (Celsius scale) in
each of the four quarters, Q_1 to Q_4, of the years 1974 to 1988 in 60 chinese
meteorologic stations; one outlier observation (YinChuan_1982) has been discarded.
The 60 stations belong to different regions in China,
which therefore define a partition of the 899 stations-year combinations.
}
\examples{
data(ChinaTemp.int)
}
\keyword{datasets}
