% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cholesterol.hist}
\alias{cholesterol.hist}
\title{Cholesterol by Gender and Age Histogram-Valued Dataset}
\format{
A data frame with 14 observations and 3 variables:
\itemize{
    \item \code{gender}: Gender (Female or Male).
    \item \code{age}: Age group (20s, 30s, 40s, 50s, 60s, 70s, 80+).
    \item \code{cholesterol}: Histogram-valued cholesterol distribution.
}
}
\source{
Billard, L. and Diday, E. (2006), Table 4.5.
}
\usage{
data(cholesterol.hist)
}
\description{
Histogram-valued cholesterol distributions for 14 gender-age groups
(7 female + 7 male age groups from 20s to 80+). Each observation has
a 10-bin histogram of cholesterol levels.
}
\examples{
data(cholesterol.hist)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis:
Conceptual Statistics and Data Mining}. Wiley, Chichester. Table 4.5.
}
\keyword{datasets}
\keyword{histogram}
\keyword{medical}
