% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_colnames.R
\name{clean_colnames}
\alias{clean_colnames}
\title{clean_colnames}
\usage{
clean_colnames(data)
}
\arguments{
\item{data}{The conventional data.}
}
\value{
Data after cleaning variable names.
}
\description{
This function is used to clean up variable names to conform to the RSDA format.
}
\examples{
data(mushroom)
mushroom.clean <- clean_colnames(data = mushroom)
}
