% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime2}
\alias{crime2}
\title{Crime Demographics Modal-Valued Dataset}
\format{
A symbolic data frame (\code{symbolic_tbl}) with 15 observations and
3 modal-valued variables:
\itemize{
    \item \code{Crime}: Modal distribution over crime types
          (violent, non-violent, none).
    \item \code{Gender}: Modal distribution over gender (male, female).
    \item \code{Age}: Modal distribution over age groups (<20, >=20).
}
}
\usage{
data(crime2)
}
\description{
Modal-valued version of the crime demographics dataset.
See \code{\link{crime}} for the wide-format version.
}
\examples{
data(crime2)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley.
}
\keyword{datasets}
\keyword{modal}
