% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{The dataSDA dataset.}
\format{
An object of class \code{data.frame} with 37 rows and 8 columns.
}
\usage{
data(dataset)
}
\description{
The summary of dataset for the dataSDA.
}
\examples{
data(dataset)
}
\keyword{datasets}
