% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flights_detail.hist}
\alias{flights_detail.hist}
\title{Airline Flights Detailed Histogram-Valued Dataset}
\format{
A data frame with 16 observations (airlines) and 5 histogram-valued
variables:
\itemize{
    \item \code{airtime}: Histogram of air time (minutes).
    \item \code{taxi_in}: Histogram of taxi-in time (minutes).
    \item \code{arrival_delay}: Histogram of arrival delay (minutes).
    \item \code{taxi_out}: Histogram of taxi-out time (minutes).
    \item \code{departure_delay}: Histogram of departure delay (minutes).
}
Row names are Airline_1 through Airline_16.
}
\source{
Billard, L. and Diday, E. (2020), Table 5-1.
}
\usage{
data(flights_detail.hist)
}
\description{
Histogram-valued dataset of 16 airlines with 5 flight performance
histograms. Each histogram has 12 bins describing the distribution
of a performance metric across flights for that airline.
}
\examples{
data(flights_detail.hist)
}
\references{
Billard, L. and Diday, E. (2020). \emph{Clustering Methodology for
Symbolic Data}. Wiley, Chichester. Table 5-1.
}
\keyword{datasets}
\keyword{flights}
\keyword{histogram}
