% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freshwater_fish.int}
\alias{freshwater_fish.int}
\title{Freshwater Fish Heavy Metal Bioaccumulation Interval Dataset}
\format{
A data frame with 12 observations and 14 variables:
\itemize{
    \item \code{body_length}: Body length (cm).
    \item \code{body_weight}: Body weight (g).
    \item \code{muscle}: Metal concentration in muscle tissue.
    \item \code{intestine}: Metal concentration in intestine.
    \item \code{stomach}: Metal concentration in stomach.
    \item \code{gills}: Metal concentration in gills.
    \item \code{liver}: Metal concentration in liver.
    \item \code{kidney}: Metal concentration in kidney.
    \item \code{liver_muscle_ratio}: Liver-to-muscle concentration ratio.
    \item \code{kidney_muscle_ratio}: Kidney-to-muscle concentration ratio.
    \item \code{gills_muscle_ratio}: Gills-to-muscle concentration ratio.
    \item \code{intestine_muscle_ratio}: Intestine-to-muscle concentration ratio.
    \item \code{stomach_muscle_ratio}: Stomach-to-muscle concentration ratio.
    \item \code{class}: Feeding category (Carnivores, Omnivores, Detritivores, Herbivores).
}
}
\source{
\url{https://github.com/Natandradesa/Kernel-Clustering-for-Interval-Data}
}
\usage{
data(freshwater_fish.int)
}
\description{
Interval-valued dataset of heavy metal concentrations in organs and tissues
of 12 freshwater fish species, grouped into 4 feeding categories (Carnivores,
Omnivores, Detritivores, Herbivores). Contains 13 interval-valued variables
measuring metal concentrations in organs and organ-to-muscle ratios.
}
\examples{
data(freshwater_fish.int)
}
\references{
Andrade, N. A., de Carvalho, F. A. T. and Pimentel, B. A. (2025).
Kernel clustering with automatic variable weighting for interval data.
\emph{Neurocomputing}, 617, 128954.
}
\keyword{clustering}
\keyword{datasets}
\keyword{interval}
