% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glucose.hist}
\alias{glucose.hist}
\title{Blood Glucose Histogram-Valued Dataset}
\format{
A data frame with 4 observations (regions) and 1 histogram-valued
variable:
\itemize{
    \item \code{glucose}: Histogram of blood glucose levels.
}
Row names are Region_1 through Region_4.
}
\source{
Billard, L. and Diday, E. (2020), Table 4-14.
}
\usage{
data(glucose.hist)
}
\description{
Histogram-valued dataset of 4 regions with a single histogram-valued
variable describing the distribution of blood glucose measurements.
}
\examples{
data(glucose.hist)
}
\references{
Billard, L. and Diday, E. (2020). \emph{Clustering Methodology for
Symbolic Data}. Wiley, Chichester. Table 4-14.
}
\keyword{datasets}
\keyword{histogram}
\keyword{medical}
