% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Health_insurance}
\alias{Health_insurance}
\title{Health Insurance Dataset}
\format{
An object of class \code{data.frame} with 51 rows and 30 columns.
}
\usage{
data(Health_insurance)
}
\description{
Health Insurance Dataset.
}
\examples{
data(Health_insurance)
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
