% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hierarchy2}
\alias{Hierarchy2}
\title{Hierarchy Dataset}
\format{
An object of class \code{data.frame} with 20 rows and 6 columns.
}
\usage{
data(Hierarchy2)
}
\description{
Hierarchy Dataset.
}
\examples{
data(Hierarchy2)
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
