% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_cor.R
\name{hist_cor}
\alias{hist_cor}
\title{The correlation of symbolic histogram variable}
\usage{
hist_cor(object, var1, var2, method)
}
\arguments{
\item{object}{A MatH object.}

\item{var1}{A symbolic histogram variable.}

\item{var2}{A symbolic histogram variable.}

\item{method}{The method to calculate correlation ex: 'BG', 'BD', 'B', 'Wass'}
}
\value{
Return covariance of the histogram variable.
}
\description{
This function compute the correlation of symbolic histogram variable.
}
\examples{
hist_cor(Blood, 'Cholesterol', 'Hemoglobin', method = 'BG')
}
