% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_mean.R
\name{hist_mean}
\alias{hist_mean}
\title{The mean of symbolic histogram variable}
\usage{
hist_mean(object, var)
}
\arguments{
\item{object}{A MatH object.}

\item{var}{A symbolic histogram variable.}
}
\value{
Return mean of the histogram variable.
}
\description{
This function compute the mean of symbolic histogram variable.
}
\examples{
hist_mean(Blood, 'Cholesterol')
}
