% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_mean_w.R
\name{hist_mean_w}
\alias{hist_mean_w}
\title{The wasserstein mean of symbolic histogram variable}
\usage{
hist_mean_w(object, var)
}
\arguments{
\item{object}{A MatH object.}

\item{var}{A symbolic histogram variable.}
}
\value{
Return wasserstein mean of the histogram variable.
}
\description{
This function compute the wasserstein mean of symbolic histogram variable.
}
\examples{
hist_mean_w(Blood, 'Cholesterol')
}
