% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_var.R
\name{hist_var}
\alias{hist_var}
\title{The variance of symbolic histogram variable}
\usage{
hist_var(object, var)
}
\arguments{
\item{object}{A MatH object.}

\item{var}{A symbolic histogram variable.}
}
\value{
Return variance of the histogram variable.
}
\description{
This function compute the variance of symbolic histogram variable.
}
\examples{
hist_var(Blood, 'Cholesterol')
}
