% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_var_w.R
\name{hist_var_w}
\alias{hist_var_w}
\title{The wasserstein variance of symbolic histogram variable}
\usage{
hist_var_w(object, var)
}
\arguments{
\item{object}{A MatH object.}

\item{var}{A symbolic histogram variable.}
}
\value{
Return wasserstein variance of the histogram variable.
}
\description{
This function compute the variance of symbolic histogram variable.
}
\examples{
hist_var_w(Blood, 'Cholesterol')
}
