% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_format_conversions.R
\name{iGAP_to_ARRAY}
\alias{iGAP_to_ARRAY}
\title{iGAP to ARRAY}
\usage{
iGAP_to_ARRAY(data, location = NULL)
}
\arguments{
\item{data}{A data.frame in iGAP format.}

\item{location}{Integer vector specifying which columns contain
comma-separated interval values.}
}
\value{
A numeric array of dimension \code{[n, p, 2]} with dimnames.
}
\description{
Convert iGAP format to a 3-dimensional array \code{[n, p, 2]}.
}
\examples{
data(abalone.iGAP)
arr <- iGAP_to_ARRAY(abalone.iGAP, 1:7)
dim(arr)
}
