% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_format_conversions.R
\name{iGAP_to_RSDA}
\alias{iGAP_to_RSDA}
\title{iGAP to RSDA}
\usage{
iGAP_to_RSDA(data, location)
}
\arguments{
\item{data}{The dataframe with the iGAP format.}

\item{location}{The location of the symbolic variable in the data.}
}
\value{
Return a symbolic_tbl dataframe with complex-encoded interval columns.
}
\description{
To convert iGAP format interval dataframe to RSDA format (symbolic_tbl).
}
\examples{
data(abalone.iGAP)
rsda <- iGAP_to_RSDA(abalone.iGAP, 1:7)
}
