% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_dist.R
\name{int_dist}
\alias{int_dist}
\title{Interval Data Distance Measures}
\usage{
int_dist(x, dist = "snEHD", gamma = 0.5, q = 1)
}
\arguments{
\item{x}{An array of dimension \code{[n, p, 2]} where \code{x[,,1]} contains the lower
bounds (min) and \code{x[,,2]} contains the upper bounds (max) for \code{n} concepts
and \code{p} interval variables.}

\item{dist}{Character string specifying the distance measure. One of:
\describe{
  \item{"GD"}{Gowda-Diday distance (Gowda & Diday, 1991)}
  \item{"IY"}{Ichino-Yaguchi distance (Ichino, 1988)}
  \item{"L1"}{L1 (midpoint) distance}
  \item{"L2"}{L2 (Euclidean midpoint) distance}
  \item{"CB"}{City-Block distance (Souza & de Carvalho, 2004)}
  \item{"HD"}{Hausdorff distance (Chavent & Lechevallier, 2002)}
  \item{"EHD"}{Euclidean Hausdorff distance}
  \item{"nEHD"}{Normalized Euclidean Hausdorff distance}
  \item{"snEHD"}{Span Normalized Euclidean Hausdorff distance}
  \item{"TD"}{Tran-Duckstein distance (Tran & Duckstein, 2002)}
  \item{"WD"}{L2-Wasserstein distance (Verde & Irpino, 2008)}
}}

\item{gamma}{Parameter for the Ichino-Yaguchi distance, \code{0 <= gamma <= 0.5}.
Default is 0.5.}

\item{q}{Parameter for the Ichino-Yaguchi distance (Minkowski exponent).
Default is 1.}
}
\value{
An object of class \code{"dist"} containing the pairwise distances.
}
\description{
Computes pairwise distance matrices for interval-valued symbolic data.
}
\examples{
# Create example interval data: 4 concepts, 3 variables
x <- array(NA, dim = c(4, 3, 2))
x[,,1] <- matrix(c(1,2,3,4, 5,6,7,8, 9,10,11,12), nrow=4)
x[,,2] <- matrix(c(3,5,6,7, 8,9,10,12, 13,15,16,18), nrow=4)
d <- int_dist(x, dist = "snEHD")
print(d)
}
\references{
Kao, C.-H. et al. (2014). Exploratory data analysis of interval-valued
  symbolic data with matrix visualization. \emph{CSDA}, 79, 14-29.

Verde, R. & Irpino, A. (2008). A new interval data distance based on the
  Wasserstein metric.
}
