% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_dist.R
\name{int_dist_all}
\alias{int_dist_all}
\title{Compute All 11 Distance Measures for Interval Data}
\usage{
int_dist_all(x, gamma = 0.5, q = 1)
}
\arguments{
\item{x}{An array of dimension \code{[n, p, 2]} of interval data.
See \code{\link{int_dist}} for details.}

\item{gamma}{Parameter for the Ichino-Yaguchi distance.
Default is 0.5.}

\item{q}{Minkowski exponent for the Ichino-Yaguchi distance.
Default is 1.}
}
\value{
A named list of \code{"dist"} objects, one per distance measure.
}
\description{
A convenience wrapper that computes all 11 distance measures
  at once for interval-valued symbolic data.
}
\examples{
x <- array(NA, dim = c(4, 3, 2))
x[,,1] <- matrix(c(1,2,3,4, 5,6,7,8, 9,10,11,12), nrow=4)
x[,,2] <- matrix(c(3,5,6,7, 8,9,10,12, 13,15,16,18), nrow=4)
all_d <- int_dist_all(x)
names(all_d)
}
\seealso{
\code{\link{int_dist}}
}
