% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_mean_estimation.R
\name{interval_mean_estimation}
\alias{interval_mean_estimation}
\title{The estimation of symbolic mean for interval variable}
\usage{
interval_mean_estimation(x)
}
\arguments{
\item{x}{A symbolic interval variable.}
}
\value{
Return mean of the interval variable.
}
\description{
This function compute the estimation of symbolic mean for interval variable.
}
\examples{
data("Mushroom")
interval_mean_estimation(Mushroom$Pileus.Cap.Width)
}
