% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_utils.R
\name{interval_utils}
\alias{interval_utils}
\title{Utility Functions for Interval Data}
\description{
Internal functions for interval data transformation.
  These are used by the exported interval statistics functions
  (\code{\link{int_mean}}, \code{\link{int_var}}, \code{\link{int_cov}},
  \code{\link{int_cor}}) and are not intended to be called directly.
}
\details{
Utility Functions for Interval Data
}
\keyword{internal}
