% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_var_estimation.R
\name{interval_var_estimation}
\alias{interval_var_estimation}
\title{The estimation of symbolic variance for interval variable}
\usage{
interval_var_estimation(x, n, alpha)
}
\arguments{
\item{x}{A symbolic interval variable.}

\item{n}{The number of elements in each group.}

\item{alpha}{The α-correction term related to the asymptotic mean of order statistics.}
}
\value{
Return mean of the interval variable.
}
\description{
This function compute the estimation of symbolic variance for interval variable.
}
\examples{
data("Mushroom")
interval_var_estimation(Mushroom$Pileus.Cap.Width, n = 3, alpha = 3/8)
}
