% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris.int}
\alias{iris.int}
\title{Iris Species Interval Dataset}
\format{
A data frame with 30 observations and 5 variables:
\itemize{
    \item \code{sepal_length}: Sepal length range (cm).
    \item \code{sepal_width}: Sepal width range (cm).
    \item \code{petal_length}: Petal length range (cm).
    \item \code{petal_width}: Petal width range (cm).
    \item \code{class}: Species (Setosa, Versicolor, Virginica).
}
}
\source{
\url{https://github.com/Natandradesa/Kernel-Clustering-for-Interval-Data}
}
\usage{
data(iris.int)
}
\description{
Interval-valued version of the classic iris dataset, aggregated from
Fisher's iris data into 30 interval observations across 3 species
(Setosa, Versicolor, Virginica). Each observation represents a group of
flowers with ranges for sepal and petal measurements.
}
\examples{
data(iris.int)
}
\references{
Andrade, N. A., de Carvalho, F. A. T. and Pimentel, B. A. (2025).
Kernel clustering with automatic variable weighting for interval data.
\emph{Neurocomputing}, 617, 128954.
}
\keyword{clustering}
\keyword{datasets}
\keyword{interval}
