% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Iris_data}
\alias{Iris_data}
\title{Iris data example}
\format{
An object of class \code{data.frame} (inherits from \code{symbolic_tbl}) with 3 rows and 4 columns.
}
\source{
\url{https://CRAN.R-project.org/package=ggESDA}
}
\usage{
data(Iris_data)
}
\description{
Iris_data interval data example.
}
\examples{
data(Iris_data)
}
\keyword{datasets}
