% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris_species.hist}
\alias{iris_species.hist}
\title{Iris Species Histogram-Valued Dataset}
\format{
A data frame with 3 observations and 5 variables:
\itemize{
    \item \code{species}: Species name (factor: Versicolor, Virginica, Setosa).
    \item \code{sepal_width}: Histogram-valued sepal width distribution.
    \item \code{sepal_length}: Histogram-valued sepal length distribution.
    \item \code{petal_width}: Histogram-valued petal width distribution.
    \item \code{petal_length}: Histogram-valued petal length distribution.
}
Row names are species names.
}
\source{
Billard, L. and Diday, E. (2020), Table 4-10.
}
\usage{
data(iris_species.hist)
}
\description{
Histogram-valued dataset of 3 iris species (Versicolor, Virginica,
Setosa) with 4 histogram-valued morphological variables and a species
label. Each histogram describes the distribution of measurements
within a species.
}
\examples{
data(iris_species.hist)
}
\references{
Billard, L. and Diday, E. (2020). \emph{Clustering Methodology for
Symbolic Data}. Wiley, Chichester. Table 4-10.
}
\keyword{datasets}
\keyword{histogram}
\keyword{iris}
