% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Loansby_risk_lvs}
\alias{Loansby_risk_lvs}
\title{Loans by risk levels: Interval Dataset}
\format{
A data frame containing 35 observations on the following 8 variables:

 \describe{
 \item{ln-inc}{The current risk category, of natural logarithm of the self-reported annual income provided by the borrower during registration.}
 \item{int-rate}{The current risk category, of the interest rate on the loan.}
 \item{open-acc}{The current risk category, of the number of open credit lines in the borrower's credit file.}
 \item{total-acc}{The current risk category, of the total number of credit lines currently in the borrower's credit file.}
 }
}
\source{
\itemize{
    \item \url{https:www.kaggle.com/wendykan/lending-club-loan-data}
    \item \url{https://CRAN.R-project.org/package=MAINT.Data}
}
}
\usage{
data(Loansby_risk_lvs)
}
\description{
This data set consist of the lower and upper bounds of the intervals for
four interval characteristics for 35 risk levels (from A1 to G5) of loans.
The original microdata is available at the Kaggle Data Science platform and
consists of 887 383 loan records characterized by 75 descriptors.
Among the large set of variables available, we focus on borrowers'
income and account and loan information aggregated by the 35 risk levels
which are considered as the units of interest.
}
\examples{
data(Loansby_risk_lvs)
}
\keyword{datasets}
