% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LoansbyPurpose.int}
\alias{LoansbyPurpose.int}
\title{Loans by purpose: Interval Dataset}
\format{
A data frame containing 14 observations on the following 4 variables:
\itemize{
    \item \code{ln-inc}: The current loan purpose of natural logarithm of the self-reported annual income provided by the borrower during registration
    \item \code{ln-revolbal}: The current loan purpose of natural logarithm of the total credit revolving balance
    \item \code{open-acc}: The current loan purpose of the number of open credit lines in the borrower's credit file
    \item \code{total-acc}: The current loan purpose, of the total number of credit lines currently in the borrower's credit file
}
}
\source{
\url{https://CRAN.R-project.org/package=MAINT.Data}
}
\usage{
data(LoansbyPurpose.int)
}
\description{
Loans by purpose interval dataset generated from LoansbyPurpose dataset.
This data set consist of the lower and upper bounds of the intervals for four
interval characteristics of the loans aggregated by their purpose.
The original microdata is available at the Kaggle Data Science platform and consists
of 887 383 loan records characterized by 75 descriptors.
Among the large set of variables available, we focus on borrowers'
income and account and loan information aggregated by the 14 loan purposes, wich are
considered as the units of interest.
}
\examples{
data(LoansbyPurpose.int)
}
\keyword{datasets}
