% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lynne1.int}
\alias{lynne1.int}
\title{Lynne1 Blood Pressure Interval Dataset}
\format{
A symbolic data frame (\code{symbolic_tbl}) with 10 observations
and 4 variables:
\itemize{
    \item \code{concept}: Character concept label.
    \item \code{Pulse Rate}: Interval-valued pulse rate (beats/min).
    \item \code{Systolic Pressure}: Interval-valued systolic pressure (mmHg).
    \item \code{Diastolic Pressure}: Interval-valued diastolic pressure (mmHg).
}
}
\source{
RSDA R package (\code{Lynne1} dataset).
}
\usage{
data(lynne1.int)
}
\description{
Interval-valued dataset of 10 observations with pulse rate, systolic
pressure, and diastolic pressure intervals.
}
\examples{
data(lynne1.int)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis: Conceptual
Statistics and Data Mining}. Wiley, Chichester.

Original data from the RSDA R package (\code{Lynne1} dataset).
}
\keyword{datasets}
\keyword{interval}
