% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_format_conversions.R
\name{MM_to_RSDA}
\alias{MM_to_RSDA}
\title{MM to RSDA}
\usage{
MM_to_RSDA(data)
}
\arguments{
\item{data}{The dataframe with the MM format (paired _min/_max columns).}
}
\value{
Return a symbolic_tbl dataframe with complex-encoded interval columns.
}
\description{
To convert MM format interval dataframe to RSDA format (symbolic_tbl).
}
\examples{
data(mushroom.int)
mm <- RSDA_to_MM(mushroom.int, RSDA = FALSE)
rsda <- MM_to_RSDA(mm)
}
