% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mtcars_data}
\alias{Mtcars_data}
\title{Mtcars data example}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}, \code{symbolic_tbl}) with 5 rows and 11 columns.
}
\source{
\url{https://CRAN.R-project.org/package=ggESDA}
}
\usage{
data(Mtcars_data)
}
\description{
mtcars.i interval and modal data example.
}
\examples{
data(Mtcars_data)
}
\keyword{datasets}
