% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupations2}
\alias{occupations2}
\title{Occupation Salaries Modal-Valued Dataset}
\format{
A symbolic data frame (\code{symbolic_tbl}) with 9 observations and 4 variables:
\itemize{
    \item \code{Occupation}: Occupation name (character).
    \item \code{Gender}: Modal distribution over gender (Male, Female).
    \item \code{Salary}: Modal distribution over 7 ordered salary bins.
    \item \code{n}: Sample size (numeric).
}
}
\usage{
data(occupations2)
}
\description{
Modal-valued version of the occupation salaries dataset.
See \code{\link{occupations}} for the wide-format version.
}
\examples{
data(occupations2)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley.
}
\keyword{datasets}
\keyword{modal}
