% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ozoneh}
\alias{Ozoneh}
\title{Complete Ozone dataset for Histogram data analysis}
\format{
a \code{MatH} istance, 1 row per station.
}
\source{
\itemize{
    \item \url{http://java.epa.gov/castnet/epa_jsp/prepackageddata.jsp}
    \item \url{ftp://ftp.epa.gov/castnet/data/metdata.zip}
    \item \url{https://CRAN.R-project.org/package=HistDAWass}
}
}
\usage{
data(Ozoneh)
}
\description{
The dataset contains MatH (matrix of histogram-valued data) object.
This data set list 84 stations located in the USA recording four variables.
Some stations contains missing data.
}
\examples{
data(Ozoneh)
}
\keyword{datasets}
