% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_format_conversions.R
\name{RSDA_to_ARRAY}
\alias{RSDA_to_ARRAY}
\title{RSDA to ARRAY}
\usage{
RSDA_to_ARRAY(data)
}
\arguments{
\item{data}{A symbolic_tbl with interval columns.}
}
\value{
A numeric array of dimension \code{[n, p, 2]} with dimnames.
Only interval (symbolic_interval) columns are included.
}
\description{
Convert RSDA format (symbolic_tbl) to a 3-dimensional array
\code{[n, p, 2]} where slice \code{[,,1]} contains the minima and
slice \code{[,,2]} contains the maxima.
}
\examples{
data(mushroom.int)
arr <- RSDA_to_ARRAY(mushroom.int)
dim(arr)  # [23, 3, 2]
}
