% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSDA_to_iGAP.R
\name{RSDA_to_iGAP}
\alias{RSDA_to_iGAP}
\title{RSDA to iGAP}
\usage{
RSDA_to_iGAP(data)
}
\arguments{
\item{data}{The RSDA format with interval dataframe.}
}
\value{
Return a dataframe with the iGAP format.
}
\description{
To convert RSDA format interval dataframe to iGAP format.
}
\examples{
data(mushroom.int)
RSDA_to_iGAP(mushroom.int)
}
