% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSDA_to_MM.R
\name{RSDA_to_MM}
\alias{RSDA_to_MM}
\title{RSDA to MM}
\usage{
RSDA_to_MM(data, RSDA)
}
\arguments{
\item{data}{The RSDA format with interval dataframe.}

\item{RSDA}{Whether to load the RSDA package.}
}
\value{
Return a dataframe with the MM format.
}
\description{
To convert RSDA format interval dataframe to MM format.
}
\examples{
data(mushroom.int)
RSDA_to_MM(mushroom.int, RSDA = FALSE)
}
