% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_utils.R
\name{set_variable_format}
\alias{set_variable_format}
\title{Set Variable Format}
\usage{
set_variable_format(data, location = NULL, var = NULL)
}
\arguments{
\item{data}{A conventional data.}

\item{location}{The location of the set variable in the data.}

\item{var}{The name of the set variable in the data.}
}
\value{
Return a dataframe in which a set variable is converted to one-hot encoding.
}
\description{
This function changes the format of the set variables
in the data to conform to the RSDA format.
}
\examples{
data("mushroom")
mushroom.set <- set_variable_format(data = mushroom, location = 8, var = "Species")
}
