% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soccer.bivar.int}
\alias{soccer.bivar.int}
\title{Soccer bivar Interval Data Set}
\format{
A data frame with 20 rows and 3 variables:
\itemize{
    \item \code{y}: The response variable Y (weight)
    \item \code{t1}: The explanatory variable T1 (height)
    \item \code{t2}: The explanatory variable T2 (age)
}
}
\source{
\url{https://CRAN.R-project.org/package=iRegression}
}
\usage{
soccer.bivar.int
}
\description{
Soccer bivar interval dataset generated from soccer.bivar dataset.
A real interval-valued data set.
}
\details{
This data set concerns the record of the Weight (Y),
Height (T1) and Age (T2) from 20 soccer teams of the premiere French championship.
}
\examples{
data(soccer.bivar.int)
}
\references{
Lima Neto, E. A., Cordeiro, G. and De Carvalho, F.A.T. (2011).
Bivariate symbolic regression models for interval-valued variables.
Journal of Statistical Computation and Simulation (Print), 81, 1727–1744.
}
\keyword{datasets}
