% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_format_conversions.R
\name{SODAS_to_ARRAY}
\alias{SODAS_to_ARRAY}
\title{SODAS to ARRAY}
\usage{
SODAS_to_ARRAY(XMLPath)
}
\arguments{
\item{XMLPath}{Disk path where the SODAS \code{*.XML} file is.}
}
\value{
A numeric array of dimension \code{[n, p, 2]} with dimnames.
}
\description{
Convert SODAS format (XML file) to a 3-dimensional array
\code{[n, p, 2]}.
}
\examples{
\dontrun{
arr <- SODAS_to_ARRAY("C:/Users/user/AppData/abalone.xml")
}
}
