% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Uscrime}
\alias{Uscrime}
\title{Us crime interval data table.}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 46 rows and 102 columns.
}
\source{
\url{https://CRAN.R-project.org/package=RSDA}
}
\usage{
data(Uscrime)
}
\description{
Us crime classic data table genetated from uscrime data.
}
\examples{
data(Uscrime)
}
\references{
Rodriguez O. (2013).
A generalization of Centre and Range method for fitting a linear
regression model to symbolic interval data using Ridge Regression, Lasso
and Elastic Net methods. The IFCS2013 conference of the International Federation of
Classification Societies, Tilburg University Holland.
}
\keyword{datasets}
