% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{video3.int}
\alias{video3.int}
\title{Video Platform User Engagement Intervals (Dataset 3)}
\format{
A symbolic data frame (\code{symbolic_tbl}) with 10 observations
and 5 interval-valued variables (V1--V5): number of visits, watches,
likes, comments, and shares.
}
\source{
GPCSIV R package (\code{video3} dataset).
}
\usage{
data(video3.int)
}
\description{
Interval-valued engagement metrics for 10 user groups on a video
platform. Variables represent ranges of visit, watch, like, comment,
and share counts.
}
\examples{
data(video3.int)
}
\references{
Makosso-Kallyth, S. and Diday, E. (2012). Adaptation of interval PCA
to symbolic histogram variables. \emph{Advances in Data Analysis and
Classification}, 6(2), 147--159.

Original data from the GPCSIV R package (\code{video3} dataset).
}
\keyword{datasets}
\keyword{interval}
