% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abalone.iGAP}
\alias{abalone.iGAP}
\title{Abalone Dataset (iGAP Format)}
\format{
An object of class \code{data.frame} with 24 rows and 7 columns.
}
\source{
UCI Machine Learning Repository.
}
\usage{
data(abalone.iGAP)
}
\description{
Interval-valued dataset of 24 units from the UCI Abalone dataset,
aggregated by sex and age. iGAP format for matrix visualization.
}
\examples{
data(abalone.iGAP)
}
\references{
Kao, C.-H. et al. (2014). Exploratory data analysis of interval-valued
symbolic data with matrix visualization. \emph{CSDA}, 79, 14-29.
}
\keyword{datasets}
\keyword{iGAP}
\keyword{interval}
