% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abalone.int}
\alias{abalone.int}
\title{Abalone Interval Dataset}
\format{
An object of class \code{data.frame} with 24 rows and 14 columns.
}
\source{
UCI Machine Learning Repository.
}
\usage{
data(abalone.int)
}
\description{
Interval-valued dataset of 24 units from the UCI Abalone dataset,
aggregated by sex and age. Standard data frame format.
}
\examples{
data(abalone.int)
}
\keyword{datasets}
\keyword{interval}
