% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acid_rain.int}
\alias{acid_rain.int}
\title{Acid Rain Pollution Indices Interval Dataset}
\format{
A data frame with 2 observations and 2 interval-valued variables:
\itemize{
    \item \code{sulphate}: Sulphate pollution index range (kg/hectares).
    \item \code{nitrate}: Nitrate pollution index range (kg/hectares).
}
}
\usage{
data(acid_rain.int)
}
\description{
Interval-valued acid rain pollution indices for sulphates and nitrates
(kg/hectares) by US state.
}
\examples{
data(acid_rain.int)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley.
Table 2.21.
}
\keyword{datasets}
\keyword{interval}
