% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_cholesterol_weight.int}
\alias{age_cholesterol_weight.int}
\title{Age-Cholesterol-Weight Interval Dataset}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7 rows and 4 columns.
}
\usage{
data(age_cholesterol_weight.int)
}
\description{
Interval-valued dataset relating age, cholesterol, and weight measurements.
}
\examples{
data(age_cholesterol_weight.int)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley.
}
\keyword{datasets}
\keyword{interval}
