% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airline_flights.hist}
\alias{airline_flights.hist}
\title{JFK Airport Airline Flights Histogram-Valued Dataset}
\format{
An object of class \code{data.frame} with 16 rows and 17 columns.
}
\usage{
data(airline_flights.hist)
}
\description{
Histogram-valued dataset of 16 airlines flying into JFK Airport.
Six variables (Flight Time, Taxi In, Arrival Delay, Taxi Out,
Departure Delay, Weather Delay) recorded as frequency distributions.
}
\examples{
data(airline_flights.hist)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley. Table 2.7.
}
\keyword{datasets}
\keyword{histogram}
