% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airline_flights}
\alias{airline_flights}
\title{Airline Flights Dataset}
\format{
An object of class \code{data.frame} with 16 rows and 17 columns.
}
\usage{
data(airline_flights)
}
\description{
Airline Flights Dataset.
}
\examples{
data(airline_flights)
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
