% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank_rates}
\alias{bank_rates}
\title{Bank Interest Rates AR Model Symbolic Dataset}
\format{
An object of class \code{data.frame} with 4 rows and 6 columns.
}
\usage{
data(bank_rates)
}
\description{
Symbolic dataset of autoregressive time series models for 4 banks.
Each bank is described by AR model order, parameters, and noise variance.
}
\examples{
data(bank_rates)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley.
Table 2.9.
}
\keyword{datasets}
\keyword{model}
\keyword{symbolic}
