% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseball.int}
\alias{baseball.int}
\title{Baseball Teams Interval Dataset}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 19 rows and 3 columns.
}
\usage{
data(baseball.int)
}
\description{
Interval-valued data for baseball teams with player statistics.
}
\examples{
data(baseball.int)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley.
}
\keyword{datasets}
\keyword{interval}
