% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseball.int}
\alias{baseball.int}
\title{Baseball Interval-Valued Dataset}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 19 rows and 3 columns.
}
\usage{
data(baseball.int)
}
\description{
Baseball Interval-Valued Dataset.
}
\examples{
data(baseball.int)
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
