% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bird.mix}
\alias{bird.mix}
\title{Bird Species Mixed Symbolic Dataset}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 20 rows and 2 columns.
}
\usage{
data(bird.mix)
}
\description{
Mixed symbolic data for bird species with interval-valued morphological
measurements and categorical symbolic variables (habitat, color).
}
\examples{
data(bird.mix)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley. Table 2.5.
}
\keyword{categorical}
\keyword{datasets}
\keyword{interval}
\keyword{mixed}
