% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bird_species_extended.mix}
\alias{bird_species_extended.mix}
\title{Bird Species Extended Mixed Symbolic Dataset}
\format{
A data frame with 3 observations and 4 symbolic variables.
}
\usage{
data(bird_species_extended.mix)
}
\description{
Three bird species (Geese, Ostrich, Penguin) with interval-valued
height, histogram-valued color distribution, and categorical
flying/migratory variables.
}
\examples{
data(bird_species_extended.mix)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley.
Table 2.19.
}
\keyword{categorical}
\keyword{datasets}
\keyword{histogram}
\keyword{interval}
\keyword{mixed}
