% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{car.int}
\alias{car.int}
\title{Car Models Interval Dataset}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8 rows and 5 columns.
}
\usage{
data(car.int)
}
\description{
Interval-valued data for car models with price, engine, speed, acceleration.
}
\examples{
data(car.int)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley.
}
\keyword{datasets}
\keyword{interval}
