% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{car_models.int}
\alias{car_models.int}
\title{Italian Car Models Interval Dataset}
\format{
A data frame with 33 observations and 9 variables:
\itemize{
    \item \code{price}: Price range (currency units).
    \item \code{engine_cc}: Engine displacement range (cc).
    \item \code{top_speed}: Top speed range (km/h).
    \item \code{acceleration}: Acceleration range (seconds 0-100 km/h).
    \item \code{wheelbase}: Wheelbase range (cm).
    \item \code{length}: Length range (cm).
    \item \code{width}: Width range (cm).
    \item \code{height}: Height range (cm).
    \item \code{class}: Car category (Utilitaria, Berlina, Ammiraglia, Sportiva).
}
}
\source{
\url{https://github.com/Natandradesa/Kernel-Clustering-for-Interval-Data}
}
\usage{
data(car_models.int)
}
\description{
Interval-valued specifications for 33 Italian car models, classified into
4 categories (Utilitaria, Berlina, Ammiraglia, Sportiva). An extended
version of the classic cars interval dataset with 8 interval-valued
variables including dimensions.
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 33 \cr
  \strong{Variables (p)} \tab 9 \cr
  \strong{Subject area} \tab Automotive \cr
  \strong{Symbolic format} \tab Interval \cr
  \strong{Analytical tasks} \tab Clustering, Classification
}
}

\examples{
data(car_models.int)
}
\references{
Andrade de Sa, J. N., Ferreira, M. R. P. and de Carvalho, F. A. T. (2025).
Kernel clustering with automatic variable weighting for interval data.
\emph{Neurocomputing}, 650, 130849.
\doi{10.1016/j.neucom.2025.130849}.
}
\keyword{clustering}
\keyword{datasets}
\keyword{interval}
