% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cardiological.int}
\alias{cardiological.int}
\title{Cardiological Examination Interval Dataset}
\format{
A data frame with 44 observations and 5 interval-valued variables:
\itemize{
    \item \code{pulse}: Pulse rate range (beats per minute).
    \item \code{systolic}: Systolic blood pressure range (mmHg).
    \item \code{diastolic}: Diastolic blood pressure range (mmHg).
    \item \code{arterial1}: First arterial measurement range.
    \item \code{arterial2}: Second arterial measurement range.
}
}
\source{
Extracted from RSDA package (\code{cardiologicalv2}).
}
\usage{
data(cardiological.int)
}
\description{
Interval-valued data from cardiological examinations of 44 patients.
Each patient is described by 5 interval-valued physiological
measurements.
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 44 \cr
  \strong{Variables (p)} \tab 5 \cr
  \strong{Subject area} \tab Medical \cr
  \strong{Symbolic format} \tab Interval \cr
  \strong{Analytical tasks} \tab Descriptive statistics, Clustering
}
}

\examples{
data(cardiological.int)
}
\references{
Rodriguez, O. (2000). Classification et modeles lineaires en analyse
des donnees symboliques. Doctoral Thesis, Universite Paris IX-Dauphine.
}
\keyword{datasets}
\keyword{interval}
