% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cars.int}
\alias{cars.int}
\title{Cars Interval Dataset}
\format{
A data frame with 27 observations and 5 variables.
}
\source{
\url{https://CRAN.R-project.org/package=MAINT.Data}
}
\usage{
data(cars.int)
}
\description{
Interval-valued data for 27 car models classified into four classes
(Utilitarian, Berlina, Sportive, Luxury), described by Price,
EngineCapacity, TopSpeed and Acceleration intervals.
}
\examples{
data(cars.int)
}
\references{
Duarte Silva, A.P., Brito, P., Filzmoser, P. and Dias, J.G. (2021).
MAINT.Data: Modelling and Analysing Interval Data in R.
\emph{R Journal}, 13(2).
}
\keyword{classification}
\keyword{datasets}
\keyword{interval}
