% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cars.int}
\alias{Cars.int}
\title{Cars Interval Dataset}
\format{
A data frame containing 27 observations on 5 variables, the first five with the interval
characteristics for 27 car models, the last one a factor indicating the model class.
}
\source{
\url{https://CRAN.R-project.org/package=MAINT.Data}
}
\usage{
data(Cars.int)
}
\description{
Cars Interval Dataset generated from Cars dataset.
This data set consist of the intervals for four characteristics
(Price, EngineCapacity, TopSpeed and Acceleration) of 27 cars models partitioned into
four different classes (Utilitarian, Berlina, Sportive and Luxury).
}
\examples{
data(Cars.int)
}
\keyword{datasets}
