% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{china_temp.int}
\alias{china_temp.int}
\title{China Meteorological Stations Quarterly Temperature Interval Dataset}
\format{
A symbolic data frame (\code{symbolic_tbl}) with 899 observations and 5 variables:
\itemize{
    \item \code{Q1}: Quarter 1 (Jan--Mar) temperature range (tenths of degrees Celsius, interval).
    \item \code{Q2}: Quarter 2 (Apr--Jun) temperature range (interval).
    \item \code{Q3}: Quarter 3 (Jul--Sep) temperature range (interval).
    \item \code{Q4}: Quarter 4 (Oct--Dec) temperature range (interval).
    \item \code{GeoReg}: Geographic region classification (factor).
}
}
\source{
\url{https://CRAN.R-project.org/package=MAINT.Data}
}
\usage{
data(china_temp.int)
}
\description{
Interval-valued temperature data (Celsius) for 60 Chinese meteorological
stations observed over the four quarters of years 1974 to 1988.
One outlier observation (YinChuan_1982) has been discarded.
}
\details{
Originates from the Long-Term Instrumental Climatic Database of the
People's Republic of China. Widely used in the SDA literature for
demonstrating standardization, clustering, self-organizing maps,
MLE and MANOVA.
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 899 \cr
  \strong{Variables (p)} \tab 5 \cr
  \strong{Subject area} \tab Climate \cr
  \strong{Symbolic format} \tab Interval \cr
  \strong{Analytical tasks} \tab Clustering
}
}

\examples{
data(china_temp.int)
}
\references{
Brito, P. and Duarte Silva, A.P. (2012). Modelling interval data with
Normal and Skew-Normal distributions. \emph{J. Appl. Stat.}, 39(1), 3-20.
}
\keyword{clustering}
\keyword{datasets}
\keyword{interval}
