% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{china_temp.int}
\alias{china_temp.int}
\title{China Meteorological Stations Quarterly Temperature Interval Dataset}
\format{
A data frame with 899 observations and 5 variables.
}
\source{
\url{https://CRAN.R-project.org/package=MAINT.Data}
}
\usage{
data(china_temp.int)
}
\description{
Interval-valued temperature data (Celsius) for 60 Chinese meteorological
stations observed over the four quarters of years 1974 to 1988.
One outlier observation (YinChuan_1982) has been discarded.
}
\details{
Originates from the Long-Term Instrumental Climatic Database of the
People's Republic of China. Widely used in the SDA literature for
demonstrating standardization, clustering, self-organizing maps,
MLE and MANOVA.
}
\examples{
data(china_temp.int)
}
\references{
Brito, P. and Duarte Silva, A.P. (2012). Modelling interval data with
Normal and Skew-Normal distributions. \emph{J. Appl. Stat.}, 39(1), 3-20.

Kao, C.-H. et al. (2014). Exploratory data analysis of interval-valued
symbolic data with matrix visualization. \emph{CSDA}, 79, 14-29.
}
\keyword{clustering}
\keyword{datasets}
\keyword{interval}
