% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime}
\alias{crime}
\title{Crime Demographics Dataset}
\format{
An object of class \code{data.frame} with 15 rows and 7 columns.
}
\usage{
data(crime)
}
\description{
Crime-related demographic variables with symbolic data types.
}
\examples{
data(crime)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley.
}
\keyword{datasets}
\keyword{symbolic}
