% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime2.modal}
\alias{crime2.modal}
\title{Crime Demographics Modal-Valued Dataset}
\format{
A symbolic data frame (\code{symbolic_tbl}) with 15 observations and
3 modal-valued variables:
\itemize{
    \item \code{Crime}: Modal distribution over crime types
          (violent, non-violent, none).
    \item \code{Gender}: Modal distribution over gender (male, female).
    \item \code{Age}: Modal distribution over age groups (<20, >=20).
}
}
\usage{
data(crime2.modal)
}
\description{
Modal-valued version of the crime demographics dataset.
See \code{\link{crime.modal}} for the wide-format version.
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 15 \cr
  \strong{Variables (p)} \tab 3 \cr
  \strong{Subject area} \tab Criminology \cr
  \strong{Symbolic format} \tab Modal \cr
  \strong{Analytical tasks} \tab Clustering, Descriptive statistics
}
}

\examples{
data(crime2.modal)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley.
}
\keyword{datasets}
\keyword{modal}
