% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime2}
\alias{crime2}
\title{Crime demographics Modal-Valued Dataset}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 15 rows and 3 columns.
}
\usage{
data(crime2)
}
\description{
Crime demographics Modal-Valued Dataset.
}
\examples{
data(crime2)
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
