% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli_routes.int}
\alias{ecoli_routes.int}
\title{E. coli Transport Routes Interval Dataset}
\format{
A symbolic data frame (\code{symbolic_tbl}) with 9 observations
(transport routes) and 5 interval-valued variables:
\itemize{
    \item \code{Y1} through \code{Y5}: Interval-valued biochemical
          pathway measurements.
}
Row names are Route_1 through Route_9.
}
\source{
Billard, L. and Diday, E. (2020), Table 8-10.
}
\usage{
data(ecoli_routes.int)
}
\description{
Interval-valued dataset of 9 E. coli transport routes with 5 interval
variables representing biochemical pathway measurements.
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 9 \cr
  \strong{Variables (p)} \tab 5 \cr
  \strong{Subject area} \tab Biology \cr
  \strong{Symbolic format} \tab Interval \cr
  \strong{Analytical tasks} \tab Clustering
}
}

\examples{
data(ecoli_routes.int)
}
\references{
Billard, L. and Diday, E. (2020). \emph{Clustering Methodology for
Symbolic Data}. Wiley, Chichester. Table 8-10.
}
\keyword{biology}
\keyword{datasets}
\keyword{interval}
