% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{employment.int}
\alias{employment.int}
\title{European Employment by Gender and Age Interval Dataset}
\format{
A data frame with 12 sex-age group observations and
interval-valued proportion variables.
}
\usage{
data(employment.int)
}
\description{
Interval-valued proportions for 12 sex-age population groups across
employment variables (employment type, education, industry sector,
occupation, marital status). Used for factorial discriminant analysis.
}
\examples{
data(employment.int)
}
\references{
Diday, E. and Noirhomme-Fraiture, M. (Eds.) (2008). \emph{Symbolic Data
Analysis and the SODAS Software}. Wiley. Table 18.1.
}
\keyword{datasets}
\keyword{discriminant}
\keyword{interval}
