% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{environment.mix}
\alias{environment.mix}
\title{EPA Environmental Data Mixed Symbolic Dataset}
\format{
A data frame with 14 observations and 17 variables. Modal-valued
columns (URBANICITY, INCOMELEVEL, EDUCATION, REGIONDEVELOPME, etc.) contain
distribution strings. Interval-valued columns (CONTROL, PM2.5, PM10, etc.)
contain complex-encoded intervals.
}
\source{
Extracted from ggESDA package (\code{Environment}).
}
\usage{
data(environment.mix)
}
\description{
Mixed symbolic dataset from the US EPA with 14 state-group observations
and 17 variables of mixed types: interval-valued environmental measurements
and modal-valued (distributional) categorical variables.
}
\examples{
data(environment.mix)
}
\references{
Sun, Y. and Billard, L. (2020). Symbolic data analysis with the
ggESDA package. \emph{Journal of Statistical Software}.
}
\keyword{datasets}
\keyword{interval}
\keyword{mixed}
\keyword{modal}
