% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{face.iGAP}
\alias{face.iGAP}
\title{Face Dataset (iGAP Format)}
\format{
An object of class \code{data.frame} with 27 rows and 6 columns.
}
\usage{
data(face.iGAP)
}
\description{
Symbolic data matrix with all interval-type variables for facial
measurements, in iGAP format.
}
\examples{
data(face.iGAP)
}
\references{
Kao, C.-H. et al. (2014). Exploratory data analysis of interval-valued
symbolic data with matrix visualization. \emph{CSDA}, 79, 14-29.
}
\keyword{datasets}
\keyword{iGAP}
\keyword{interval}
