% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Face.iGAP}
\alias{Face.iGAP}
\title{Face iGAP format Dataset}
\format{
An object of class \code{data.frame} with 27 rows and 6 columns.
}
\usage{
data(Face.iGAP)
}
\description{
Symbolic data matrix with all the variables of interval type.
}
\examples{
data(Face.iGAP)
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
