% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finance.int}
\alias{finance.int}
\title{Finance Sector Interval Dataset}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 14 rows and 7 columns.
}
\usage{
data(finance.int)
}
\description{
Interval-valued data for 14 business sectors described by job-related
financial variables (job cost codes, activity codes, budgets).
Used for PCA demonstrations.
}
\examples{
data(finance.int)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley. Table 5.2.
}
\keyword{PCA}
\keyword{datasets}
\keyword{interval}
