% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fuel_consumption}
\alias{Fuel_consumption}
\title{Fuel Consumption Dataset}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 3 columns.
}
\usage{
data(Fuel_consumption)
}
\description{
Fuel Consumption Dataset.
}
\examples{
data(Fuel_consumption)
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
