% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fungi.int}
\alias{fungi.int}
\title{Fungi Morphological Measurements Interval Dataset}
\format{
A data frame with 55 observations and 6 variables:
\itemize{
    \item \code{pileus_width}: Width of the pileus (cap).
    \item \code{stipe_width}: Width of the stipe (stem).
    \item \code{stipe_thickness}: Thickness of the stipe.
    \item \code{spore_height}: Height of the spores.
    \item \code{spore_width}: Width of the spores.
    \item \code{class}: Fungus genus (Amanita, Agaricus, Boletus).
}
}
\source{
\url{https://github.com/Natandradesa/Kernel-Clustering-for-Interval-Data}
}
\usage{
data(fungi.int)
}
\description{
Interval-valued morphological measurements for 55 fungi specimens from
3 genera (Amanita, Agaricus, Boletus). Contains 5 interval-valued variables
describing pileus and stipe dimensions and spore characteristics.
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 55 \cr
  \strong{Variables (p)} \tab 6 \cr
  \strong{Subject area} \tab Biology \cr
  \strong{Symbolic format} \tab Interval \cr
  \strong{Analytical tasks} \tab Clustering
}
}

\examples{
data(fungi.int)
}
\references{
Andrade de Sa, J. N., Ferreira, M. R. P. and de Carvalho, F. A. T. (2025).
Kernel clustering with automatic variable weighting for interval data.
\emph{Neurocomputing}, 650, 130849.
\doi{10.1016/j.neucom.2025.130849}.
}
\keyword{clustering}
\keyword{datasets}
\keyword{interval}
