% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{health_insurance.mix}
\alias{health_insurance.mix}
\title{Health Insurance Mixed Symbolic Dataset}
\format{
An object of class \code{data.frame} with 51 rows and 30 columns.
}
\usage{
data(health_insurance.mix)
}
\description{
Health insurance data grouped by disease type and gender with
classical and symbolic variables of mixed types.
}
\examples{
data(health_insurance.mix)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley. Tables 2.1-2.2.
}
\keyword{datasets}
\keyword{mixed}
\keyword{symbolic}
