% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{health_insurance2}
\alias{health_insurance2}
\title{Health Insurance Modal-Valued Dataset}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 6 rows and 6 columns.
}
\usage{
data(health_insurance2)
}
\description{
Modal-valued version of the health insurance dataset.
}
\examples{
data(health_insurance2)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley. Table 2.2b.
}
\keyword{datasets}
\keyword{modal}
