% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{health_insurance2}
\alias{health_insurance2}
\title{Health Insurance Modal-Valued Dataset}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 6 rows and 6 columns.
}
\usage{
data(health_insurance2)
}
\description{
Health Insurance Modal-Valued Dataset.
}
\examples{
data(health_insurance2)
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
