% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hierarchy.int}
\alias{hierarchy.int}
\title{Hierarchy Interval Dataset}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 20 rows and 6 columns.
}
\usage{
data(hierarchy.int)
}
\description{
Interval-valued version of the hierarchy dataset.
}
\examples{
data(hierarchy.int)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley. Table 2.15.
}
\keyword{datasets}
\keyword{interval}
