% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hierarchy}
\alias{hierarchy}
\title{Hierarchy Dataset}
\format{
An object of class \code{data.frame} with 20 rows and 6 columns.
}
\usage{
data(hierarchy)
}
\description{
Classical dataset illustrating hierarchical data structures.
}
\examples{
data(hierarchy)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley. Table 2.15.
}
\keyword{datasets}
\keyword{hierarchical}
