% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{horses.int}
\alias{horses.int}
\title{Horse Breeds Interval Dataset}
\format{
A data frame with 8 observations and 6 interval-valued variables.
}
\usage{
data(horses.int)
}
\description{
Interval-valued data for 8 horse breeds (CES, CMA, PEN, TES, CEN,
LES, PES, PAM) described by 6 variables: minimum/maximum weight,
minimum/maximum height, cost of mares, cost of fillies.
}
\details{
Extensively used in SDA for demonstrating divisive clustering, distance
computation, hierarchy/pyramid construction, and complete objects.
}
\examples{
data(horses.int)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley. Table 7.14.
}
\keyword{clustering}
\keyword{datasets}
\keyword{interval}
