% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGAP_to_MM.R
\name{iGAP_to_MM}
\alias{iGAP_to_MM}
\title{iGAP to MM}
\usage{
iGAP_to_MM(data, location)
}
\arguments{
\item{data}{The iGAP file.}

\item{location}{The location of the symbolic variable in the data.}
}
\value{
A CSV data file.
}
\description{
To convert iGAP files to CSV files.
}
\examples{
data(Abalone.iGAP)
Abalone <- iGAP_to_MM(Abalone.iGAP, c(1, 2, 3, 4, 5, 6, 7))
}
