% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{joggers.mix}
\alias{joggers.mix}
\title{Joggers Mixed Symbolic Dataset}
\format{
A symbolic data frame (\code{symbolic_tbl}) with 10 observations
(jogger groups) and 2 variables:
\itemize{
    \item \code{pulse_rate}: Interval-valued resting pulse rate range (bpm).
    \item \code{running_time}: Histogram-valued distribution of running
          times (minutes).
}
Row names are Group_1 through Group_10.
}
\source{
Billard, L. and Diday, E. (2020), Table 2-5.
}
\usage{
data(joggers.mix)
}
\description{
Mixed symbolic dataset of 10 jogger groups with one interval-valued
variable (pulse rate) and one histogram-valued variable (running time
distribution).
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 10 \cr
  \strong{Variables (p)} \tab 2 \cr
  \strong{Subject area} \tab Sports \cr
  \strong{Symbolic format} \tab Mixed (interval, histogram) \cr
  \strong{Analytical tasks} \tab Clustering
}
}

\examples{
data(joggers.mix)
}
\references{
Billard, L. and Diday, E. (2020). \emph{Clustering Methodology for
Symbolic Data}. Wiley, Chichester. Table 2-5.
}
\keyword{datasets}
\keyword{histogram}
\keyword{interval}
\keyword{mixed}
