% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{judge2.int}
\alias{judge2.int}
\title{Judge 2 Interval-Valued Ratings}
\format{
A symbolic data frame (\code{symbolic_tbl}) with 6 observations
and 4 interval-valued variables (V1--V4).
}
\source{
Simulated data distributed as \code{GPCSIV::Judge2}.
}
\usage{
data(judge2.int)
}
\description{
Simulated interval-valued agricultural-product grades assigned by Judge 2
for 6 regions on 4 variables, distributed with the GPCSIV package.
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 6 \cr
  \strong{Variables (p)} \tab 4 \cr
  \strong{Subject area} \tab Methodology \cr
  \strong{Symbolic format} \tab Interval \cr
  \strong{Analytical tasks} \tab PCA
}
}

\examples{
data(judge2.int)
}
\references{
Makosso-Kallyth, S. (2013). Analysis of m sets of symbolic interval
variables. \emph{Revue des Nouvelles Technologies de l'Information},
RNTI-E-25, 97--108.
}
\keyword{datasets}
\keyword{interval}
