% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{judge3.int}
\alias{judge3.int}
\title{Judge 3 Interval-Valued Ratings}
\format{
A symbolic data frame (\code{symbolic_tbl}) with 6 observations
and 4 interval-valued variables (V1--V4).
}
\source{
GPCSIV R package (\code{Judge3} dataset).
}
\usage{
data(judge3.int)
}
\description{
Interval-valued ratings from Judge 3 for 6 regions on 4 variables.
From a study of generalized principal component analysis for
interval-valued data (GPCSIV).
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 6 \cr
  \strong{Variables (p)} \tab 4 \cr
  \strong{Subject area} \tab Methodology \cr
  \strong{Symbolic format} \tab Interval \cr
  \strong{Analytical tasks} \tab PCA
}
}

\examples{
data(judge3.int)
}
\references{
Makosso-Kallyth, S. and Diday, E. (2012). Adaptation of interval PCA
to symbolic histogram variables. \emph{Advances in Data Analysis and
Classification}, 6(2), 147--159.

Original data from the GPCSIV R package (\code{Judge3} dataset).
}
\keyword{datasets}
\keyword{interval}
