% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loans_by_purpose.int}
\alias{loans_by_purpose.int}
\title{Loans by Purpose Interval Dataset}
\format{
A data frame with 14 observations and 4 interval-valued variables:
\itemize{
    \item \code{ln_inc}: Natural logarithm of self-reported annual income.
    \item \code{ln_revolbal}: Natural logarithm of total credit revolving balance.
    \item \code{open_acc}: Number of open credit lines.
    \item \code{total_acc}: Total number of credit lines.
}
}
\source{
\url{https://CRAN.R-project.org/package=MAINT.Data}
}
\usage{
data(loans_by_purpose.int)
}
\description{
Interval-valued data for loan characteristics aggregated by their purpose.
Original microdata contains 887,383 loan records from Kaggle.
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 14 \cr
  \strong{Variables (p)} \tab 4 \cr
  \strong{Subject area} \tab Finance \cr
  \strong{Symbolic format} \tab Interval \cr
  \strong{Analytical tasks} \tab Descriptive statistics, Clustering
}
}

\examples{
data(loans_by_purpose.int)
}
\keyword{datasets}
\keyword{interval}
