% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lung_cancer.hist}
\alias{lung_cancer.hist}
\title{Lung Cancer Treatments by State Histogram-Valued Dataset}
\format{
An object of class \code{data.frame} with 2 rows and 2 columns.
}
\usage{
data(lung_cancer.hist)
}
\description{
Histogram-valued distribution of lung cancer treatment counts by US state.
}
\examples{
data(lung_cancer.hist)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley.
Table 2.20.
}
\keyword{datasets}
\keyword{histogram}
