% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mushroom.int}
\alias{mushroom.int}
\title{Mushroom Species Interval Dataset}
\format{
A data frame with 23 observations and interval-valued variables.
}
\usage{
data(mushroom.int)
}
\description{
Interval-valued version of the mushroom dataset. See \code{\link{mushroom}}.
}
\examples{
data(mushroom.int)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley. Table 3.2.
}
\keyword{datasets}
\keyword{interval}
