% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mushroom}
\alias{mushroom}
\title{Mushroom Data Set}
\format{
A data frame with 23 observations and 5 variables named Species, Pileus Cap
Width, Stipe Length, Stipe Thickness, and Edibility.

\itemize{
    \item \code{Species}: The class of mushroom.
    \item \code{Pileus Cap Width}: The pileus cap width of the mushroom.
    \item \code{Stipe Length}: The stipe length of the mushroom.
    \item \code{Stipe Thickness}: The stipe thickness of the mushroom.
    \item \code{Edibility}: The edibility of mushroom (U: unknown, Y:  Yes, N: No, T: Toxic).
}
}
\source{
Billard, L. and Diday, E. (2006) Symbolic Data Analysis: Conceptual Statistics and Data Mining John Wiley & Sons, Ltd.
}
\usage{
data(mushroom)
}
\description{
The mushroom data set consists of a set of 23 species described by 3 interval variables.
These mushroom species are members of the genus Agaricies.
The specific variables and their values are extracted from the Fungi of California Species.
}
\examples{
data(mushroom)
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
