% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mushroom_fuzzy}
\alias{mushroom_fuzzy}
\title{Mushroom Species Fuzzy/Symbolic Dataset}
\format{
An object of class \code{data.frame} with 4 rows and 9 columns.
}
\usage{
data(mushroom_fuzzy)
}
\description{
Extended mushroom data with fuzzy stipe thickness (Small/Average/Large),
numerical stipe length, interval cap size, and categorical cap colour
for two Amanita species.
}
\examples{
data(mushroom_fuzzy)
}
\references{
Diday, E. and Noirhomme-Fraiture, M. (Eds.) (2008). \emph{Symbolic Data
Analysis and the SODAS Software}. Wiley. Tables 1.14-1.16.
}
\keyword{datasets}
\keyword{fuzzy}
\keyword{symbolic}
