% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nycflights.int}
\alias{nycflights.int}
\title{New York City Flights Interval Dataset}
\format{
A list containing FlightsDF, FlightsUnits, and FlightsIdt.
}
\source{
\url{https://CRAN.R-project.org/package=MAINT.Data}
}
\usage{
data(nycflights.int)
}
\description{
Interval-valued dataset with 142 units and four interval-valued variables
from the nycflights13 package, aggregated by month and carrier.
}
\examples{
data(nycflights.int)
}
\references{
Duarte Silva, A.P., Brito, P., Filzmoser, P. and Dias, J.G. (2021).
MAINT.Data: Modelling and Analysing Interval Data in R. \emph{R Journal}, 13(2).
}
\keyword{datasets}
\keyword{interval}
