% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupations}
\alias{occupations}
\title{Occupation Salaries Dataset}
\format{
An object of class \code{data.frame} with 9 rows and 11 columns.
}
\usage{
data(occupations)
}
\description{
Salary ranges for different occupations.
}
\examples{
data(occupations)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley.
}
\keyword{datasets}
\keyword{interval}
