% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polish_voivodships.int}
\alias{polish_voivodships.int}
\title{Polish Voivodships Socio-Economic Intervals}
\format{
A symbolic data frame (\code{symbolic_tbl}) with 18 observations
(voivodships) and 9 interval-valued variables:
\itemize{
    \item \code{V1} through \code{V9}: Interval-valued socio-economic
          indicators aggregated across counties within each voivodship.
}
Row names are voivodship names (e.g., Dolnoslaskie, Lubelskie).
}
\source{
clusterSim R package (\code{data_pathtinger} dataset).
}
\usage{
data(polish_voivodships.int)
}
\description{
Interval-valued dataset of 18 Polish voivodships (administrative regions)
with 9 socio-economic interval variables describing demographic and
economic characteristics at the county (powiat) level.
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 18 \cr
  \strong{Variables (p)} \tab 9 \cr
  \strong{Subject area} \tab Socioeconomics \cr
  \strong{Symbolic format} \tab Interval \cr
  \strong{Analytical tasks} \tab Clustering
}
}

\examples{
data(polish_voivodships.int)
}
\references{
Dudek, A. and Pelka, M. (2022). \emph{symbolicDA: Analysis of Symbolic
Data}. R package.

Walesiak, M. and Dudek, A. (2020). \emph{clusterSim: Searching for
Optimal Clustering Procedure for a Data Set}. R package.
}
\keyword{datasets}
\keyword{interval}
\keyword{socioeconomic}
