% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{profession.int}
\alias{profession.int}
\title{Profession Work Salary Time Interval-Valued Dataset}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 15 rows and 4 columns.
}
\usage{
data(profession.int)
}
\description{
Profession Work Salary Time Interval-Valued Dataset.
}
\examples{
data(profession.int)
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
