% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_data.R
\name{search_data}
\alias{search_data}
\title{Search and filter datasets that align with the specified options.}
\usage{
search_data(...)
}
\arguments{
\item{...}{task, type, n, p, tag}
}
\value{
Return a list with the the specified options of dataset.
}
\description{
Show the different types of data sets in this package.
}
\examples{
search_data(n > 100, p < 10)
}
