% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated.hist}
\alias{simulated.hist}
\title{Simulated Histogram-Valued Dataset}
\format{
A data frame with 5 observations and 2 histogram-valued variables:
\itemize{
    \item \code{Y1}: Histogram-valued variable 1.
    \item \code{Y2}: Histogram-valued variable 2.
}
Row names are Obs_1 through Obs_5.
}
\source{
Billard, L. and Diday, E. (2020), Table 7-26.
}
\usage{
data(simulated.hist)
}
\description{
Small simulated histogram-valued dataset of 5 observations with 2
histogram-valued variables. Useful for testing and demonstrating
histogram-valued statistical methods.
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 5 \cr
  \strong{Variables (p)} \tab 2 \cr
  \strong{Subject area} \tab Methodology \cr
  \strong{Symbolic format} \tab Histogram \cr
  \strong{Analytical tasks} \tab Clustering
}
}

\examples{
data(simulated.hist)
}
\references{
Billard, L. and Diday, E. (2020). \emph{Clustering Methodology for
Symbolic Data}. Wiley, Chichester. Table 7-26.
}
\keyword{datasets}
\keyword{histogram}
\keyword{simulated}
