% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soccer_bivar.int}
\alias{soccer_bivar.int}
\title{French Soccer Championship Bivariate Interval Dataset}
\format{
A data frame with 20 rows and 3 interval-valued variables:
\itemize{
    \item \code{y}: Weight (response variable, kg).
    \item \code{t1}: Height (explanatory variable, cm).
    \item \code{t2}: Age (explanatory variable, years).
}
}
\source{
\url{https://CRAN.R-project.org/package=iRegression}
}
\usage{
data(soccer_bivar.int)
}
\description{
Interval-valued data for 20 teams from the French premier soccer championship.
Contains ranges of Weight (response), Height and Age (explanatory variables).
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 20 \cr
  \strong{Variables (p)} \tab 3 \cr
  \strong{Subject area} \tab Sports \cr
  \strong{Symbolic format} \tab Interval \cr
  \strong{Analytical tasks} \tab Regression
}
}

\examples{
data(soccer_bivar.int)
}
\references{
Lima Neto, E. A., Cordeiro, G. and De Carvalho, F.A.T. (2011).
Bivariate symbolic regression models for interval-valued variables.
\emph{Journal of Statistical Computation and Simulation}, 81, 1727-1744.
}
\keyword{datasets}
\keyword{interval}
\keyword{regression}
