% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SODAS_to_iGAP.R
\name{SODAS_to_iGAP}
\alias{SODAS_to_iGAP}
\title{SODAS to iGAP}
\usage{
SODAS_to_iGAP(XMLPath)
}
\arguments{
\item{XMLPath}{Disk path where the SODAS *.XML file is.}
}
\value{
Return a dataframe with the iGAP format.
}
\description{
To convert SODAS format interval dataframe to the iGAP format.
}
\examples{
## Not run:
data(Abalone)
}
