% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SODAS_to_MM.R
\name{SODAS_to_MM}
\alias{SODAS_to_MM}
\title{SODAS to MM}
\usage{
SODAS_to_MM(XMLPath)
}
\arguments{
\item{XMLPath}{Disk path where the SODAS *.XML file is.}
}
\value{
Return a dataframe with the MM format.
}
\description{
To convert SODAS format interval dataframe to the MM format.
}
\examples{
## Not run:
data(Abalone)
}
