% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_income.hist}
\alias{state_income.hist}
\title{State Income Histogram-Valued Dataset}
\format{
A data frame with 6 observations (states) and 4 histogram-valued
variables:
\itemize{
    \item \code{Y1} through \code{Y4}: Histogram-valued income distribution
          variables.
}
Row names are State_1 through State_6.
}
\source{
Billard, L. and Diday, E. (2020), Table 7-18.
}
\usage{
data(state_income.hist)
}
\description{
Histogram-valued dataset of 6 US states with 4 income distribution
histograms. Each histogram describes the distribution of household
income within a state.
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 6 \cr
  \strong{Variables (p)} \tab 4 \cr
  \strong{Subject area} \tab Economics \cr
  \strong{Symbolic format} \tab Histogram \cr
  \strong{Analytical tasks} \tab Clustering
}
}

\examples{
data(state_income.hist)
}
\references{
Billard, L. and Diday, E. (2020). \emph{Clustering Methodology for
Symbolic Data}. Wiley, Chichester. Table 7-18.
}
\keyword{datasets}
\keyword{histogram}
\keyword{income}
