% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_clusters.int}
\alias{synthetic_clusters.int}
\title{Synthetic Interval Clusters Dataset}
\format{
A symbolic data frame (\code{symbolic_tbl}) with 125 observations and 7 variables:
\itemize{
    \item \code{V1} through \code{V6}: Six interval-valued variables.
    \item \code{class}: Cluster membership (1--5, set-valued).
}
}
\source{
Extracted from symbolicDA package (\code{data_symbolic}).
}
\usage{
data(synthetic_clusters.int)
}
\description{
Synthetic interval-valued dataset with 125 observations in 5 groups of
25 each, described by 6 interval-valued variables and a cluster label.
Designed for benchmarking interval data clustering algorithms.
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 125 \cr
  \strong{Variables (p)} \tab 7 \cr
  \strong{Subject area} \tab Methodology \cr
  \strong{Symbolic format} \tab Interval \cr
  \strong{Analytical tasks} \tab Clustering
}
}

\examples{
data(synthetic_clusters.int)
}
\references{
Dudek, A. and Pelka, M. (2022). \emph{symbolicDA}: Analysis of Symbolic
Data. R package.
}
\keyword{clustering}
\keyword{datasets}
\keyword{interval}
\keyword{synthetic}
