% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_clusters.int}
\alias{synthetic_clusters.int}
\title{Synthetic Interval Clusters Dataset}
\format{
A data frame with 125 observations and 7 variables:
\itemize{
    \item \code{X1} through \code{X6}: Six interval-valued variables.
    \item \code{cluster}: Cluster membership (1-5).
}
}
\source{
Extracted from symbolicDA package (\code{data_symbolic}).
}
\usage{
data(synthetic_clusters.int)
}
\description{
Synthetic interval-valued dataset with 125 observations in 5 groups of
25 each, described by 6 interval-valued variables and a cluster label.
Designed for benchmarking interval data clustering algorithms.
}
\examples{
data(synthetic_clusters.int)
}
\references{
Dudek, A. and Pelka, M. (2022). \emph{symbolicDA}: Analysis of Symbolic
Data. R package.
}
\keyword{clustering}
\keyword{datasets}
\keyword{interval}
\keyword{synthetic}
