% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uscrime.int}
\alias{uscrime.int}
\title{US Crime Statistics Interval Dataset}
\format{
A data frame with 46 observations and 102 interval-valued variables.
}
\source{
Extracted from RSDA package (\code{uscrime_int}).
}
\usage{
data(uscrime.int)
}
\description{
Interval-valued crime statistics for 46 US states, containing 102
interval-valued variables covering various crime types and rates.
Originally from the RSDA package.
}
\examples{
data(uscrime.int)
}
\references{
Rodriguez, O. (2000). Classification et modeles lineaires en analyse
des donnees symboliques. Doctoral Thesis, Universite Paris IX-Dauphine.
}
\keyword{crime}
\keyword{datasets}
\keyword{interval}
