% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{veterinary.int}
\alias{veterinary.int}
\title{Veterinary Interval Dataset}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 3 columns.
}
\usage{
data(veterinary.int)
}
\description{
Interval-valued veterinary dataset with animal measurements.
}
\examples{
data(veterinary.int)
}
\references{
Billard, L. and Diday, E. (2006). \emph{Symbolic Data Analysis}. Wiley.
}
\keyword{datasets}
\keyword{interval}
