% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{video3.int}
\alias{video3.int}
\title{Video Platform User Engagement Intervals (Dataset 3)}
\format{
A symbolic data frame (\code{symbolic_tbl}) with 10 observations
and 5 interval-valued variables (V1--V5): number of visits, watches,
likes, comments, and shares.
}
\source{
Simulated data distributed as \code{GPCSIV::video3}.
}
\usage{
data(video3.int)
}
\description{
Simulated interval-valued engagement metrics for 10 user groups on a video
platform. Variables represent ranges of visit, watch, like, comment,
and share counts.
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 10 \cr
  \strong{Variables (p)} \tab 5 \cr
  \strong{Subject area} \tab Digital media \cr
  \strong{Symbolic format} \tab Interval \cr
  \strong{Analytical tasks} \tab PCA
}
}

\examples{
data(video3.int)
}
\references{
Makosso-Kallyth, S. (2013). Analysis of m sets of symbolic interval
variables. \emph{Revue des Nouvelles Technologies de l'Information},
RNTI-E-25, 97--108.
}
\keyword{datasets}
\keyword{interval}
