% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine.int}
\alias{wine.int}
\title{Wine Chemical Properties Interval Dataset}
\format{
A data frame with 33 observations and 10 variables:
\itemize{
    \item \code{V1} through \code{V9}: Nine interval-valued chemical/physical
          property measurements.
    \item \code{class}: Wine group (1 or 2).
}
}
\source{
\url{https://github.com/Natandradesa/Kernel-Clustering-for-Interval-Data}
}
\usage{
data(wine.int)
}
\description{
Interval-valued chemical and physical properties of 33 wine samples
classified into 2 groups. Contains 9 interval-valued measurement variables.
Used as a benchmark for interval data clustering algorithms.
}
\section{Metadata}{

\tabular{ll}{
  \strong{Sample size (n)} \tab 33 \cr
  \strong{Variables (p)} \tab 10 \cr
  \strong{Subject area} \tab Food science \cr
  \strong{Symbolic format} \tab Interval \cr
  \strong{Analytical tasks} \tab Clustering
}
}

\examples{
data(wine.int)
}
\references{
Andrade, N. A., de Carvalho, F. A. T. and Pimentel, B. A. (2025).
Kernel clustering with automatic variable weighting for interval data.
\emph{Neurocomputing}, 617, 128954.
}
\keyword{clustering}
\keyword{datasets}
\keyword{interval}
