% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_cup.int}
\alias{world_cup.int}
\title{World Cup Soccer Teams Interval Dataset}
\format{
A data frame with 2 observations and 5 variables.
}
\usage{
data(world_cup.int)
}
\description{
Interval-valued data for soccer teams grouped by World Cup qualification
status. Includes age, weight, height ranges and covariance.
}
\examples{
data(world_cup.int)
}
\references{
Diday, E. and Noirhomme-Fraiture, M. (Eds.) (2008). \emph{Symbolic Data
Analysis and the SODAS Software}. Wiley. Table 1.9, p.13.
}
\keyword{datasets}
\keyword{interval}
